refactor: 로그아웃 API 네트워크 모듈로 마이그레이션
