+++ data/account/src/main/java/com/into/websoso/data/account/datasource/AccountRemoteDataSource.kt
@@ -11,5 +11,10 @@ interface AccountRemoteDataSource {
         authToken: AuthToken,
     ): AccountEntity
 
+    suspend fun postLogout(
+        refreshToken: String,
+        deviceIdentifier: String,
+    )
+
     suspend fun postReissue(refreshToken: String): TokenEntity
 }
